requirements.txt
python-dotenv
discord.py
psycopg2-binary
python-binary-memcached
